login transaction timer & modal
on the login page add a new component that starts a timer when the login widget renders to monitor the idx transaction. check the timestamp every 5 seconds to see if the idx transaction has expired. if it expires show a modal with a button to clear and restart the transaction